fixed bugs, cleaned up code, added slider
fixed the bug where if user click any button whiling the program is
running, the output is not desired, by disabling the buttons when the
program is running. Added slider for users to adjust the numbers of
columns and rows. The slider still has bugs, which will be updated in
the future